+++ packages/orm/src/relations/hasMany.test.ts
@@ -1697,6 +1697,7 @@ describe('hasMany through', () => {
       columns = this.setColumns((t) => ({
         postId: t.integer().foreignKey(() => Post, 'id'),
         tagId: t.integer().foreignKey(() => Tag, 'id'),
+        ...t.primaryKey(['postId', 'tagId']),
       }));
 
       relations = {