+++ packages/pqb/src/db.test.ts
@@ -9,6 +9,7 @@ import {
 } from './test-utils/test-utils';
 import { createDb } from './db';
 import { columnTypes } from './columnSchema';
+import { QueryLogger } from './queryMethods';
 
 describe('db', () => {
   useTestDatabase();
@@ -26,12 +27,30 @@ describe('db', () => {
     );
   });
 
+  describe('primaryKeys', () => {
+    it('should collect primary keys from schema', () => {
+      const table = db('table', (t) => ({
+        id: t.serial().primaryKey(),
+        name: t.text().primaryKey(),
+      }));
+      expect(table.primaryKeys).toEqual(['id', 'name']);
+    });
+
+    it('should set primary keys from primaryKey in schema', () => {
+      const table = db('table', (t) => ({
+        ...t.primaryKey(['id', 'name']),
+      }));
+      expect(table.primaryKeys).toEqual(['id', 'name']);
+    });
+  });
+
   describe('overriding column types', () => {
     it('should return date as string by default', async () => {
       await User.create(userData);
 
       const db = createDb({ adapter, columnTypes });
       const table = db('user', (t) => ({
+        id: t.serial().primaryKey(),
         createdAt: t.timestamp(),
       }));
 
@@ -47,12 +66,15 @@ describe('db', () => {
       const db = createDb({
         adapter,
         columnTypes: {
+          serial: columnTypes.serial,
           timestamp() {
             return columnTypes.timestamp().parse((input) => new Date(input));
           },
         },
       });
+
       const table = db('user', (t) => ({
+        id: t.serial().primaryKey(),
         createdAt: t.timestamp(),
       }));
 
@@ -69,6 +91,66 @@ describe('db', () => {
 
       const table = db('table');
       expect(table.query.autoPreparedStatements).toBe(false);
+    });
+  });
+
+  describe('noPrimaryKey', () => {
+    it('should throw error when no primary key by default', () => {
+      const db = createDb({ adapter, columnTypes });
+
+      expect(() =>
+        db('table', (t) => ({
+          name: t.text(),
+        })),
+      ).toThrow(`Table table has no primary key`);
+    });
+
+    it('should throw error when no primary key when noPrimaryKey is set to `error`', () => {
+      const db = createDb({ adapter, columnTypes, noPrimaryKey: 'error' });
+
+      expect(() =>
+        db('table', (t) => ({
+          name: t.text(),
+        })),
+      ).toThrow(`Table table has no primary key`);
+    });
+
+    it('should not throw when no column shape is provided', () => {
+      const db = createDb({ adapter, columnTypes });
+
+      expect(() => db('table')).not.toThrow();
+    });
+
+    it('should warn when no primary key and noPrimaryKey is set to `warning`', () => {
+      const logger = { warn: jest.fn() };
+      const db = createDb({
+        adapter,
+        columnTypes,
+        noPrimaryKey: 'warning',
+        logger: logger as unknown as QueryLogger,
+      });
+
+      db('table', (t) => ({
+        name: t.text(),
+      }));
+
+      expect(logger.warn).toBeCalledWith('Table table has no primary key');
+    });
+
+    it('should do nothing when no primary key and noPrimaryKey is set to `ignore`', () => {
+      const logger = { warn: jest.fn() };
+      const db = createDb({
+        adapter,
+        columnTypes,
+        noPrimaryKey: 'ignore',
+        logger: logger as unknown as QueryLogger,
+      });
+
+      db('table', (t) => ({
+        name: t.text(),
+      }));
+
+      expect(logger.warn).not.toBeCalled();
     });
   });
 });