+++ packages/pqb/src/queryMethods/log.ts
@@ -11,6 +11,7 @@ export type QueryLogObject = {
 
 export type QueryLogger = {
   log(message: string): void;
+  warn(message: string): void;
   error(message: string): void;
 };