+++ packages/pqb/CHANGELOG.md
@@ -1,5 +1,11 @@
 # pqb
 
+## 0.7.1
+
+### Patch Changes
+
+- Handle tables without primary key
+
 ## 0.7.0
 
 ### Minor Changes